feat: extract togglePlay (328 lines) + handleAnalyze (296 lines) to store actions (E2+E3)
Created src/stores/actions/togglePlay.js — handles audio playback,
cache check, TTS generation with Gemini, Web Audio setup, polling for
in-flight requests. Receives audioRef and isTogglingPlay via
dependency injection.

Created src/stores/actions/analyzePoem.js — handles insight generation,
cache check, streaming Gemini responses, insight parsing, DB translation
save-back. Reads/writes poemStore + useModalStore for toast.

app.jsx reduced from ~3000 to 2278 lines (~720 lines extracted).
Both functions replaced with 1-line thin wrappers.

406 tests pass.
