+++ src/app.jsx
@@ -33,6 +33,8 @@ import { useAudioStore } from './stores/audioStore';
 import { useUIStore } from './stores/uiStore';
 import { useModalStore } from './stores/modalStore';
 import { fetchPoem as fetchPoemAction } from './stores/actions/fetchPoem';
+import { togglePlay as togglePlayAction } from './stores/actions/togglePlay';
+import { analyzePoem as analyzePoemAction } from './stores/actions/analyzePoem';
 import { getRecentSeenIds, markPoemSeen, pruneSeenPoems } from './utils/seenPoems.js';
 import { transliterate } from './utils/transliterate.js';
 import { filterPoemsByCategory } from './utils/filterPoems.js';
@@ -711,631 +713,9 @@ export default function DiwanApp() {
     </div>
   );
 
-  const togglePlay = async () => {
-    if (isTogglingPlay.current) {
-      addLog('Audio', 'Play toggle already in progress — skipping', 'info');
-      return;
-    }
-    isTogglingPlay.current = true;
-    addLog(
-      'UI Event',
-      `🎵 Play button clicked | Poem: ${current?.poet} - ${current?.title} | ID: ${current?.id}`,
-      'info'
-    );
-    track('audio_play', { poet: current?.poet });
+  const togglePlay = () => togglePlayAction({ audioRef, isTogglingPlay, current, addLog, track });
 
-    if (isPlaying) {
-      audioRef.current.pause();
-      setIsPlaying(false);
-      track('audio_pause', { poet: current?.poet });
-      addLog('UI Event', '⏸️ Pause button clicked', 'info');
-      isTogglingPlay.current = false;
-      return;
-    }
-
-    if (audioUrl) {
-      try {
-        await audioRef.current.play();
-        setIsPlaying(true);
-      } catch (e) {
-        addLog('Audio', 'Playback failed, resetting audio URL', 'info');
-        setAudioUrl(null);
-      }
-      isTogglingPlay.current = false;
-      return;
-    }
-
-    // iOS Safari autoplay policy: audio.play() must be called synchronously within a user
-    // gesture handler. Calling it here (before any await/setTimeout) "unlocks" the audio
-    // element so that deferred playback — after prefetch polling or API generation — is not
-    // blocked by the browser's autoplay restrictions. Without this unlock, audio silently
-    // fails to play after a TTS prefetch or Gemini generation completes asynchronously.
-    // We mute first to prevent a brief audible blip if a previous src is still loaded on
-    // the element, then immediately pause and restore muted to its original value.
-    if (audioRef.current) {
-      const wasMuted = audioRef.current.muted;
-      audioRef.current.muted = true;
-      const unlockPlay = audioRef.current.play();
-      if (unlockPlay !== undefined) {
-        unlockPlay.catch(() => {}); // silence errors from empty/invalid src
-      }
-      audioRef.current.pause();
-      audioRef.current.muted = wasMuted;
-    }
-
-    // Set loading state FIRST (before duplicate check) for better UX
-    setIsGeneratingAudio(true);
-
-    const doGenerate = async () => {
-      // CHECK CACHE FIRST
-      if (FEATURES.caching && current?.id) {
-        const cacheStart = performance.now();
-        const cached = await cacheOperations.get(CACHE_CONFIG.stores.audio, current.id);
-        const cacheTime = performance.now() - cacheStart;
-
-        if (cached?.blob) {
-          const sizeMB = (cached.blob.size / (1024 * 1024)).toFixed(2);
-          addLog(
-            'Audio Cache',
-            `✓ Cache HIT (${cacheTime.toFixed(0)}ms)${cached.metadata?.model ? ` | Model: ${cached.metadata.model}` : ''} | Size: ${sizeMB}MB | Instant playback`,
-            'success'
-          );
-          useUIStore.getState().incrementCacheStat('audioHits');
-
-          const u = URL.createObjectURL(cached.blob);
-          setAudioUrl(u);
-          audioRef.current.src = u;
-          audioRef.current.load();
-          audioRef.current
-            .play()
-            .then(() => setIsPlaying(true))
-            .catch((err) => {
-              if (FEATURES.logging) console.warn('[Audio] Playback failed:', err.message);
-              addLog('Audio', `Cached playback failed: ${err.message}`, 'error');
-            });
-          setIsGeneratingAudio(false); // Clear loading state
-          isTogglingPlay.current = false;
-          return;
-        } else {
-          addLog(
-            'Audio Cache',
-            `✗ Cache MISS (${cacheTime.toFixed(0)}ms) | Generating from API...`,
-            'info'
-          );
-          useUIStore.getState().incrementCacheStat('audioMisses');
-        }
-      }
-
-      // Mark request as in-flight
-      usePoemStore.getState().addActiveAudio(current?.id);
-
-      const ttsContent = getTTSContent(current);
-
-      // Calculate request metrics
-      const requestBody = JSON.stringify({
-        contents: [{ parts: [{ text: ttsContent }] }],
-        generationConfig: {
-          responseModalities: TTS_CONFIG.responseModalities,
-          speechConfig: {
-            voiceConfig: {
-              prebuiltVoiceConfig: { voiceName: TTS_CONFIG.voiceName },
-            },
-          },
-        },
-      });
-      const requestSize = new Blob([requestBody]).size;
-      const estimatedTokens = Math.ceil(ttsContent.length / 4);
-      const arabicTextChars = current?.arabic?.length || 0;
-
-      addLog(
-        'Audio API',
-        `→ Starting generation | Model: ${API_MODELS.tts} | Request: ${(requestSize / 1024).toFixed(1)}KB | ${arabicTextChars} chars Arabic | Est. ${estimatedTokens} tokens`,
-        'info'
-      );
-
-      setAudioError(null);
-
-      try {
-        const apiStart = performance.now();
-        const url = `${apiUrl}/api/ai/${API_MODELS.tts}/generateContent`;
-        const fetchOptions = {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: requestBody,
-        };
-        const { res, model: ttsModel } = await fetchTTSWithFallback(url, fetchOptions, {
-          addLog,
-          label: 'Audio API',
-        });
-
-        if (!res.ok) {
-          const errorText = await res.text();
-          addLog(
-            'Audio API Error',
-            `[${ttsModel}] HTTP ${res.status}: ${errorText.substring(0, 200)}`,
-            'error'
-          );
-          if (res.status === 429) {
-            setAudioError(
-              'Recitation temporarily unavailable — too many requests. Please wait a moment and try again.'
-            );
-            throw new Error('Rate limited (429)');
-          }
-          throw new Error(`API returned ${res.status}: ${res.statusText}`);
-        }
-
-        const data = await res.json();
-        const apiTime = performance.now() - apiStart;
-
-        if (!data.candidates || data.candidates.length === 0) {
-          addLog(
-            'Audio API Error',
-            `[${ttsModel}] No candidates in response. Full response: ${JSON.stringify(data).substring(0, 300)}`,
-            'error'
-          );
-          throw new Error('Recitation failed - no audio candidates returned');
-        }
-
-        const b64 = data.candidates?.[0]?.content?.parts?.[0]?.inlineData?.data;
-        if (b64) {
-          const conversionStart = performance.now();
-          const blob = pcm16ToWav(b64);
-          const conversionTime = performance.now() - conversionStart;
-
-          if (blob) {
-            // Calculate audio metrics
-            const audioSizeMB = (blob.size / (1024 * 1024)).toFixed(2);
-            const audioSizeKB = (blob.size / 1024).toFixed(1);
-            // Estimate audio duration from PCM samples (24kHz, 16-bit, mono)
-            const pcmBytes = atob(b64.replace(/\s/g, '')).length;
-            const samples = pcmBytes / 2; // 16-bit = 2 bytes per sample
-            const audioDuration = samples / 24000; // 24kHz sample rate
-            const tokensPerSecond = (estimatedTokens / (apiTime / 1000)).toFixed(1);
-            const totalTime = apiTime + conversionTime;
-
-            addLog(
-              'Audio API',
-              `✓ [${ttsModel}] Complete | API: ${(apiTime / 1000).toFixed(2)}s | Convert: ${conversionTime.toFixed(0)}ms | Total: ${(totalTime / 1000).toFixed(2)}s`,
-              'success'
-            );
-            addLog(
-              'Audio Metrics',
-              `[${ttsModel}] Audio: ${audioDuration.toFixed(1)}s | Size: ${audioSizeKB}KB (${audioSizeMB}MB) | Speed: ${tokensPerSecond} tok/s`,
-              'success'
-            );
-
-            const u = URL.createObjectURL(blob);
-            setAudioUrl(u);
-            audioRef.current.src = u;
-            audioRef.current.load();
-            audioRef.current
-              .play()
-              .then(() => setIsPlaying(true))
-              .catch((err) => {
-                if (FEATURES.logging) console.warn('[Audio] Playback failed:', err.message);
-                addLog('Audio', `Playback failed: ${err.message}`, 'error');
-              });
-
-            // CACHE THE AUDIO BLOB
-            if (FEATURES.caching && current?.id) {
-              const cacheStart = performance.now();
-              await cacheOperations.set(CACHE_CONFIG.stores.audio, current.id, {
-                blob,
-                metadata: {
-                  poet: current.poet,
-                  title: current.title,
-                  size: blob.size,
-                  duration: audioDuration,
-                  model: ttsModel,
-                },
-              });
-              const cacheTime = performance.now() - cacheStart;
-              addLog(
-                'Audio Cache',
-                `Audio cached for future playback (${cacheTime.toFixed(0)}ms) | Saves ${(apiTime / 1000).toFixed(1)}s on replay`,
-                'success'
-              );
-            }
-          }
-        }
-      } catch (e) {
-        Sentry.captureException(e);
-        addLog('Audio System Error', `${e.message} | Poem ID: ${current?.id}`, 'error');
-        track('audio_error', { error: (e.message || '').slice(0, 100) });
-        setIsPlaying(false);
-      } finally {
-        setIsGeneratingAudio(false);
-        usePoemStore.getState().removeActiveAudio(current?.id); // Clean up in-flight tracking
-        isTogglingPlay.current = false;
-      }
-    };
-
-    // Check if request already in flight - poll until it completes
-    if (usePoemStore.getState().hasActiveAudio(current?.id)) {
-      addLog('Audio', `Audio generation already in progress - waiting for completion`, 'info');
-
-      // Poll every 500ms to check if the request completed
-      const pollInterval = setInterval(async () => {
-        if (!usePoemStore.getState().hasActiveAudio(current?.id)) {
-          clearInterval(pollInterval);
-          usePoemStore.getState().removePollingInterval(pollInterval);
-
-          // Request completed - check cache and play
-          const cached = await cacheOperations.get(CACHE_CONFIG.stores.audio, current.id);
-          if (cached?.blob) {
-            addLog(
-              'Audio',
-              `✓ Background audio generation completed${cached.metadata?.model ? ` [${cached.metadata.model}]` : ''} - playing from cache`,
-              'success'
-            );
-            const u = URL.createObjectURL(cached.blob);
-            setAudioUrl(u);
-            audioRef.current.src = u;
-            audioRef.current.load();
-            audioRef.current
-              .play()
-              .then(() => setIsPlaying(true))
-              .catch((err) => {
-                if (FEATURES.logging) console.warn('[Audio] Playback failed:', err.message);
-                addLog('Audio', `Playback failed: ${err.message}`, 'error');
-              });
-          } else {
-            addLog(
-              'Audio',
-              'Prefetch failed — retrying audio generation automatically...',
-              'error'
-            );
-            if (!isTogglingPlay.current) {
-              isTogglingPlay.current = true;
-              await doGenerate();
-            }
-            return;
-          }
-          setIsGeneratingAudio(false);
-        }
-      }, 500);
-
-      usePoemStore.getState().addPollingInterval(pollInterval);
-
-      // Safety timeout - clear after 60 seconds (some large poems take 40+ seconds)
-      setTimeout(() => {
-        clearInterval(pollInterval);
-        usePoemStore.getState().removePollingInterval(pollInterval);
-        if (usePoemStore.getState().hasActiveAudio(current?.id)) {
-          addLog(
-            'Audio',
-            `Audio generation taking longer than expected - checking one more time...`,
-            'info'
-          );
-
-          // Final check before giving up
-          setTimeout(async () => {
-            const finalCheck = await cacheOperations.get(CACHE_CONFIG.stores.audio, current.id);
-            if (finalCheck?.blob) {
-              addLog('Audio', `✓ Audio completed after extended wait - playing now`, 'success');
-              const u = URL.createObjectURL(finalCheck.blob);
-              setAudioUrl(u);
-              audioRef.current.src = u;
-              audioRef.current.load();
-              audioRef.current
-                .play()
-                .then(() => setIsPlaying(true))
-                .catch((err) => {
-                  if (FEATURES.logging) console.warn('[Audio] Playback failed:', err.message);
-                  addLog('Audio', `Playback failed: ${err.message}`, 'error');
-                });
-            } else {
-              addLog('Audio', `Audio generation timeout - please try again`, 'error');
-            }
-            usePoemStore.getState().removeActiveAudio(current?.id);
-            setIsGeneratingAudio(false);
-          }, 10000); // Wait 10 more seconds for slow API
-        }
-      }, 60000);
-
-      isTogglingPlay.current = false;
-      return;
-    }
-
-    await doGenerate();
-  };
-
-  const handleAnalyze = async () => {
-    addLog(
-      'UI Event',
-      `🔍 Dive In button clicked | Poem: ${current?.poet} - ${current?.title} | ID: ${current?.id}`,
-      'info'
-    );
-
-    if (interpretation || isInterpreting) return;
-    track('insight_requested', { poet: current?.poet });
-
-    // Set loading state FIRST (before duplicate check) for better UX
-    setIsInterpreting(true);
-
-    // Check if request already in flight - poll until it completes
-    if (usePoemStore.getState().hasActiveInsight(current?.id)) {
-      addLog(
-        'Insights',
-        `Insights generation already in progress - waiting for completion`,
-        'info'
-      );
-
-      // Poll every 500ms to check if the request completed
-      const pollInterval = setInterval(async () => {
-        if (!usePoemStore.getState().hasActiveInsight(current?.id)) {
-          clearInterval(pollInterval);
-          usePoemStore.getState().removePollingInterval(pollInterval);
-
-          // Request completed - check cache and display
-          const cached = await cacheOperations.get(CACHE_CONFIG.stores.insights, current.id);
-          if (cached?.interpretation) {
-            addLog(
-              'Insights',
-              `✓ Background insights generation completed - displaying results`,
-              'success'
-            );
-            setInterpretation(cached.interpretation);
-          } else {
-            addLog('Insights', `Background insights generation failed - retrying`, 'info');
-            // Retry the request
-            setTimeout(() => handleAnalyze(), 100);
-            return;
-          }
-          setIsInterpreting(false);
-        }
-      }, 500);
-
-      usePoemStore.getState().addPollingInterval(pollInterval);
-
-      // Safety timeout - clear after 60 seconds (some insights take time)
-      setTimeout(() => {
-        clearInterval(pollInterval);
-        usePoemStore.getState().removePollingInterval(pollInterval);
-        if (usePoemStore.getState().hasActiveInsight(current?.id)) {
-          addLog(
-            'Insights',
-            `Insights generation taking longer than expected - checking one more time...`,
-            'info'
-          );
-
-          // Final check before giving up
-          setTimeout(async () => {
-            const finalCheck = await cacheOperations.get(CACHE_CONFIG.stores.insights, current.id);
-            if (finalCheck?.interpretation) {
-              addLog(
-                'Insights',
-                `✓ Insights completed after extended wait - displaying now`,
-                'success'
-              );
-              setInterpretation(finalCheck.interpretation);
-            } else {
-              addLog('Insights', `Insights generation timeout - please try again`, 'error');
-            }
-            usePoemStore.getState().removeActiveInsight(current?.id);
-            setIsInterpreting(false);
-          }, 10000); // Wait 10 more seconds for slow API
-        }
-      }, 60000);
-
-      return;
-    }
-
-    // Mark request as in-flight
-    usePoemStore.getState().addActiveInsight(current?.id);
-
-    // CHECK CACHE FIRST
-    if (FEATURES.caching && current?.id) {
-      const cacheStart = performance.now();
-      const cached = await cacheOperations.get(CACHE_CONFIG.stores.insights, current.id);
-      const cacheTime = performance.now() - cacheStart;
-
-      if (cached?.interpretation) {
-        const charCount = cached.interpretation.length;
-        const estTokens = Math.ceil(charCount / 4);
-        addLog(
-          'Insights Cache',
-          `✓ Cache HIT (${cacheTime.toFixed(0)}ms) | ${charCount} chars (≈${estTokens} tokens) | Instant load`,
-          'success'
-        );
-        useUIStore.getState().incrementCacheStat('insightsHits');
-        setInterpretation(cached.interpretation);
-        setIsInterpreting(false); // Clear loading state
-        usePoemStore.getState().removeActiveInsight(current?.id); // Clean up tracking
-        return;
-      } else {
-        addLog(
-          'Insights Cache',
-          `✗ Cache MISS (${cacheTime.toFixed(0)}ms) | Generating from API...`,
-          'info'
-        );
-        useUIStore.getState().incrementCacheStat('insightsMisses');
-      }
-    }
-
-    let insightText = '';
-    let apiStartTime = null;
-
-    try {
-      // Use streaming if feature flag is enabled
-      if (FEATURES.streaming) {
-        const poetInfo = current?.poet ? ` by ${current.poet}` : '';
-        const promptText = `Deep Analysis of${poetInfo}:\n\n${current?.arabic}`;
-        const requestSize = new Blob([
-          JSON.stringify({ contents: [{ parts: [{ text: promptText }] }] }),
-        ]).size;
-        const estimatedInputTokens = Math.ceil(
-          (promptText.length + INSIGHTS_SYSTEM_PROMPT.length) / 4
-        );
-        const promptChars = promptText.length;
-        const arabicTextChars = current?.arabic?.length || 0;
-        const systemPromptChars = INSIGHTS_SYSTEM_PROMPT.length;
-
-        addLog(
-          'Insights API',
-          `→ Starting streaming | Request: ${(requestSize / 1024).toFixed(1)}KB | ${promptChars} chars (${arabicTextChars} Arabic + ${systemPromptChars} system) | Est. ${estimatedInputTokens} tokens`,
-          'info'
-        );
-
-        setInterpretation(''); // Clear previous interpretation
-        apiStartTime = performance.now();
-        const apiStart = apiStartTime;
-        let firstChunkTime = null;
-        let chunkCount = 0;
-        let totalTime = 0;
-
-        const insightsStreamBody = JSON.stringify({
-          contents: [{ parts: [{ text: promptText }] }],
-          systemInstruction: { parts: [{ text: INSIGHTS_SYSTEM_PROMPT }] },
-        });
-        const res = await geminiTextFetch(
-          'streamGenerateContent',
-          insightsStreamBody,
-          'Insights failed',
-          addLog
-        );
-
-        const reader = res.body.getReader();
-        const decoder = new TextDecoder();
-        let accumulatedText = '';
-        let buffer = '';
-
-        while (true) {
-          const { done, value } = await reader.read();
-          if (done) break;
-
-          // Decode chunk and add to buffer
-          buffer += decoder.decode(value, { stream: true });
-
-          // Process complete lines from buffer
-          const lines = buffer.split('\n');
-          buffer = lines.pop() || ''; // Keep incomplete line in buffer
-
-          for (const line of lines) {
-            if (line.startsWith('data: ')) {
-              const jsonStr = line.slice(6).trim();
-              if (jsonStr === '[DONE]') continue;
-
-              try {
-                const data = JSON.parse(jsonStr);
-                const text = data.candidates?.[0]?.content?.parts?.[0]?.text || '';
-                if (text) {
-                  if (!firstChunkTime) {
-                    firstChunkTime = performance.now() - apiStart;
-                    addLog(
-                      'Insights API',
-                      `← First chunk received (${firstChunkTime.toFixed(0)}ms) | Streaming...`,
-                      'info'
-                    );
-                  }
-                  chunkCount++;
-                  accumulatedText += text;
-                  setInterpretation(accumulatedText); // Real-time UI update
-                }
-              } catch (parseError) {
-                // Skip malformed JSON chunks
-                console.debug('Skipping malformed chunk:', jsonStr);
-              }
-            }
-          }
-        }
-
-        insightText = accumulatedText;
-        totalTime = performance.now() - apiStart;
-        const charCount = insightText.length;
-        const estimatedTokens = Math.ceil(charCount / 4);
-        const tokensPerSecond = (estimatedTokens / (totalTime / 1000)).toFixed(1);
-        const avgChunkSize = charCount / chunkCount;
-
-        addLog(
-          'Insights API',
-          `✓ Streaming complete | Total: ${(totalTime / 1000).toFixed(2)}s | TTFT: ${(firstChunkTime / 1000).toFixed(2)}s | ${chunkCount} chunks`,
-          'success'
-        );
-        addLog(
-          'Insights Metrics',
-          `${charCount} chars (≈${estimatedTokens} tokens) | ${tokensPerSecond} tok/s | Avg chunk: ${avgChunkSize.toFixed(0)} chars`,
-          'success'
-        );
-      } else {
-        // Non-streaming fallback (original implementation)
-        addLog('Insights', 'Analyzing poem...', 'info');
-        const poetInfoFallback = current?.poet ? ` by ${current.poet}` : '';
-        const insightsFallbackBody = JSON.stringify({
-          contents: [
-            { parts: [{ text: `Deep Analysis of${poetInfoFallback}:\n\n${current?.arabic}` }] },
-          ],
-          systemInstruction: { parts: [{ text: INSIGHTS_SYSTEM_PROMPT }] },
-        });
-        const res = await geminiTextFetch(
-          'generateContent',
-          insightsFallbackBody,
-          'Insights failed',
-          addLog
-        );
-        const data = await res.json();
-        insightText = data.candidates?.[0]?.content?.parts?.[0]?.text;
-        setInterpretation(insightText);
-        addLog('Insights', 'Analysis complete', 'success');
-      }
-
-      // CACHE THE INSIGHTS
-      if (FEATURES.caching && current?.id && insightText) {
-        const cacheStart = performance.now();
-        await cacheOperations.set(CACHE_CONFIG.stores.insights, current.id, {
-          interpretation: insightText,
-          metadata: {
-            poet: current.poet,
-            title: current.title,
-            charCount: insightText.length,
-            tokens: Math.ceil(insightText.length / 4),
-          },
-        });
-        const cacheTime = performance.now() - cacheStart;
-        const elapsedTime = apiStartTime
-          ? ((performance.now() - apiStartTime) / 1000).toFixed(1)
-          : '2-8';
-        addLog(
-          'Insights Cache',
-          `Insights cached for future use (${cacheTime.toFixed(0)}ms) | Saves ${elapsedTime}s on reload`,
-          'success'
-        );
-      }
-
-      // Save translation back to database for future visitors (fire-and-forget)
-      if (current?.isFromDatabase && current?.id && insightText) {
-        const parts = parseInsight(insightText);
-        if (parts?.poeticTranslation) {
-          saveTranslation(current.id, {
-            translation: parts.poeticTranslation.replace(/\n/g, '*'),
-            explanation: parts.depth || null,
-            authorBio: parts.author || null,
-          });
-        }
-      }
-
-      track('insight_completed', {
-        poet: current?.poet,
-        cached: !!(FEATURES.caching && current?.id && insightText),
-      });
-
-      // Flash gold check sparkle on the sidebar icon
-      if (insightText) {
-        useModalStore.getState().showToastTimed('insight', 1500);
-      }
-    } catch (e) {
-      Sentry.captureException(e);
-      addLog('Analysis Error', `${e.message} | Poem ID: ${current?.id}`, 'error');
-      track('insight_error', { error: (e.message || '').slice(0, 100) });
-      // Show partial results if streaming was interrupted
-      if (FEATURES.streaming && insightText) {
-        addLog('Insights', 'Showing partial results', 'warning');
-      }
-    } finally {
-      setIsInterpreting(false);
-      usePoemStore.getState().removeActiveInsight(current?.id); // Clean up in-flight tracking
-    }
-  };
+  const handleAnalyze = () => analyzePoemAction({ current, addLog, track });
 
   const handleFetch = () => fetchPoemAction({ addLog, track, emitEvent, navigate, markPoemSeen });
